add debug, clean up istio commands
